+++ app/assets/stylesheets/application.scss
@@ -22,6 +22,14 @@
 a {
     text-decoration: none;
 }
+button{
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+    outline: none;
+    padding: 0;
+    appearance: none;
+}
 .wrapper{
     background-color: #0073ff48;
     width: 100%;
