+++ app/controllers/admin/tweets_controller.rb
@@ -3,4 +3,14 @@ class Admin::TweetsController < ApplicationController
     def index
         @tweets = Tweet.all.page(params[:page]).per(30)
     end
+
+    def show
+        @tweet = Tweet.find(params[:id])
+    end
+
+    def destroy
+        tweet = Tweet.find(params[:id])
+        tweet.destroy
+        redirect_to admin_tweets_path
+    end
 end
