+++ app/javascript/application.js
@@ -1,3 +1,11 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
-import "@hotwired/turbo-rails"
+//= require jquery3
+//= require jquery_ujs
+//= require popper
+//= require bootstrap
 import "controllers"
+import jquery from "jquery"
+window.$ = jquery
+$(function() {
+  console.log("Hello Rails7!");
+})
