+++ app/views/admin/homes/top.html.erb
@@ -26,7 +26,7 @@
                 <h2>サ活投稿一覧</h2>
                 <span class="mt-1">不適切な投稿は管理者が削除することができます。</span>
                 <div class="mx-auto">
-                    <%=link_to "投稿一覧画面へ", admin_saunas_path, class: "btn btn-primary mt-4"%>
+                    <%=link_to "投稿一覧画面へ", admin_tweets_path, class: "btn btn-primary mt-4"%>
                 </div>
             </div>
