+++ app/views/admin/saunas/index.html.erb
@@ -2,23 +2,23 @@
 
 <div class="container">
     <div class="row">
-        <div class="col-md-1"></div>
-        <div class="col-md-11">
+        <div class="col-10 offset-1">
             <div class="row mt-5">
-                <h1>サウナ施設一覧</h1>
+                <h1 class="text-center">サウナ施設一覧</h1>
             </div>
 
             <div class="row mt-5">
-                <div class="col-md-10 mt-3">
+                <div class="col-12 mt-3">
                     <table class="table table-striped table-hover">
                         <tr>
-                            <th>#</th>   <th>サウナ施設</th> <th>住所</th>
+                            <th>#</th>   <th>サウナ施設</th> <th>住所</th> <th></th>
                         </tr>
                         <% @saunas.each do |sauna|%>
                             <tr onclick='window.location="<%= admin_sauna_path(sauna.id)%>"' role="link">
                                 <td><%=sauna.id%></td>
                                 <td><%=sauna.name%></td>
                                 <td><%=sauna.address%></td>
+                                <td class="text-center"><%=button_to "削除ボタン！", admin_sauna_path(sauna.id), method: :delete, data: {confirm: "この操作は取り消せません。注意してください。"}, class: "btn btn-danger"%></td>
                             </tr>
                         <%end%>
                     </table>
