+++ app/views/admin/saunas/new.html.erb
@@ -2,20 +2,19 @@
 
 <div class="container">
     <div class="row">
-        <div class="col-md-1"></div>
-        <div class="col-md-11">
+        <div class="col-10 offset-1">
             <div class="row mt-5">
-                <h1>サウナ施設の投稿</h1>
+                <h1 class="text-center">サウナ施設の投稿</h1>
             </div>
-            <div class="row mt-3">
-                <div class="col-auto alert alert-danger">
+            <div class="row mt-3 d-flex justify-content-center">
+                <div class="col-10 alert alert-danger">
                 現在、サウナ室の温度と水風呂の温度が入力されていない場合でも施設を作成できてしまうバグが発生中です。<br>
                 必要項目すべてに回答したかどうか確認した上で投稿ボタンを押してください。
                 </div>
             </div>
 
-            <div class="row mt-5">
-                <%=render 'admin/saunas/form', sauna: @sauna, sauna_infos: @sauna_infos, water: @water%>
+            <div class="row mt-3">
+                <%=render 'admin/saunas/form', sauna: @sauna, sauna_infos: @sauna_infos, water: @water, admin: current_manager%>
             </div>
 
         </div>
