+++ app/views/admin/saunas/show.html.erb
@@ -62,7 +62,7 @@
                     <%=link_to '編集する', edit_admin_sauna_path(@sauna.id), class: "btn btn-secondary"%>
                 </div>
                 <div class="col-md-auto">
-                    <%=button_to "削除する", admin_sauna_path(@sauna.id), method: :delete, data: { turbo_confirm: "このサウナ施設を削除します。この操作は取り消せません。" }, class: "btn btn-danger"%>
+                    <%=button_to "削除する", admin_sauna_path(@sauna.id), method: :delete, data: {confirm: "このサウナ施設を削除します。この操作は取り消せません。" }, class: "btn btn-danger"%>
                 </div>
             </div>
