+++ app/views/admin/sessions/new.html.erb
@@ -2,34 +2,29 @@
 
 <div class="container">
   <div class="row">
-    <div class="col-md-1"></div>
-    <div class="col-md-11">
-        <div class="row mt-5"><h1>管理者ログイン</h1></div>
-        <div class="row">
+    <div class="col-10 offset-1">
+        <div class="row mt-5"><h1 class="text-center">管理者ログイン</h1></div>
+        <div class="row mt-2">
           <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
-            <div class="col-md-4 field mt-5">
+            <div class="col-6 offset-3 field mt-5">
               <%= f.label :email, "管理者用メールアドレス" %><br />
               <%= f.email_field :email, autofocus: true, autocomplete: "email", class: "form-control" %>
             </div>
 
-            <div class="col-md-4 field mt-4">
+            <div class="col-6 offset-3 field mt-4">
               <%= f.label :password, "パスワード" %><br />
               <%= f.password_field :password, autocomplete: "current-password", class: "form-control" %>
             </div>
 
-            <% if devise_mapping.rememberable? %>
-              <div class="field mt-4">
-                <%= f.check_box :remember_me %>
-                <%= f.label :remember_me, "次回から自動ログイン" %>
-              </div>
-            <% end %>
+            <div class="col-6 offset-3 field mt-4">
+              <%= f.label :name, "名前" %><br />
+              <%= f.text_field :name, autocomplete: "name", class: "form-control" %>
+            </div>
 
-            <div class="actions mt-5">
-              <%= f.submit "管理者としてログインする", class: "btn btn-primary" %>
+            <div class="actions text-center mt-5">
+              <%= f.submit "管理者としてログインする", class: "btn btn-primary col-6" %>
             </div>
           <% end %>
-
-          <div class="col mt-5"><%= render "admin/shared/links" %></div>
         </div>
     </div>
   </div>
