+++ app/views/admin/tweets/show.html.erb
@@ -2,12 +2,22 @@
 
 <div class="container">
     <div class="row">
-        <div class="col-md-1"></div>
-        <div class="col-md-11">
+        <div class="col-10 offset-1">
             <div class="row mt-5">
                 <h1>サ活投稿詳細</h1>
                 <p>ユーザーの投稿を確認することができます。</p>
                 <p>ここから削除することが可能です。削除する際はよく確認してください。</p>
+            </div>
+
+            <div class="row mt-5">
+                <h3 class="pb-2"><%=@tweet.user.name%></h3>の投稿
+
+                <p class="text-right ps-2"><%=@tweet.comment%></p>
+            </div>
+
+            <div class="row mt-5">
+                <h3>これは削除ボタンです。気をつけてください。</h3>
+                <%=link_to "削除する", admin_tweet_path(@tweet.id), method: :delete, data: {confirm: "本当に削除しますか？この操作は取り消せません。"}, class: "btn btn-danger"%>
             </div>
         </div>
     </div>
