+++ app/views/public/favorites/_btn_sauna.html.erb
@@ -1,12 +1,12 @@
 <% if sauna.sauna_favorited_by?(current_user) %>
     <span class="fs-5 ps-2">
-    <%= link_to sauna_sauna_favorite_path(sauna.id), data: {"turbo-method": :delete} do %>
+    <%= button_to sauna_sauna_favorite_path(sauna.id), method: :delete, remote: true do %>
         <i class="fa-regular fa-thumbs-up fa-lg" style="color: #e83030;"></i> <%= sauna.sauna_favorites.count %>
     <% end %>
     </span>
     <% else %>
     <span class="fs-5 ps-2">
-    <%= link_to sauna_sauna_favorite_path(sauna.id), data: {"turbo-method": :post} do %>
+    <%= button_to sauna_sauna_favorite_path(sauna.id), method: :post, remote: true do %>
         <i class="fa-regular fa-thumbs-up fa-lg" style="color: #141619;"></i> <%= sauna.sauna_favorites.count %>
     <% end %>
     </span>
