+++ app/views/public/favorites/_btn_tweet.html.erb
@@ -1,12 +1,12 @@
 <% if tweet.favorited_by?(current_user) %>
     <span class="fs-5 ps-2">
-    <%= link_to tweet_favorite_path(tweet.id), data: {"turbo-method": :delete} do %>
+    <%= link_to tweet_favorite_path(tweet.id), method: :delete, remote: true do %>
         <i class="fa-regular fa-thumbs-up fa-lg" style="color: #e83030;"></i> <%= tweet.favorites.count %>
     <% end %>
     </span>
     <% else %>
     <span class="fs-5 ps-2">
-    <%= link_to tweet_favorite_path(tweet.id), data: {"turbo-method": :post} do %>
+    <%= link_to tweet_favorite_path(tweet.id), method: :post, remote: true do %>
         <i class="fa-regular fa-thumbs-up fa-lg" style="color: #141619;"></i> <%= tweet.favorites.count %>
     <% end %>
     </span>
