+++ app/views/public/sessions/new.html.erb
@@ -8,7 +8,7 @@
       </div>
 
       <% if flash[:notice]%>
-        <div class="row mt-5"> 
+        <div class="row mt-5 d-flex justify-content-center"> 
           <div class="col-auto">
             <p class = "alert alert-warning"><%= flash[:notice] %></p>
           </div>
