+++ app/views/public/shared/_error_messages.html.erb
@@ -1,12 +1,12 @@
 <% if resource.errors.any? %>
   <div id="error_explanation" data-turbo-cache="false">
-    <h2>
+    <h3 class="text-center">
       <%= I18n.t("errors.messages.not_saved",
                  count: resource.errors.count,
                  resource: resource.class.model_name.human.downcase)
        %>
-    </h2>
-    <ul>
+    </h3>
+    <ul class="text-center mb-0">
       <% resource.errors.full_messages.each do |message| %>
         <li><%= message %></li>
       <% end %>
