+++ app/views/public/users/unsubscribe.html.erb
@@ -21,7 +21,7 @@
 
             <div class="row mt-5">
                 <div class="col-8 offset-2">
-                    <%= button_to "退会する", withdrawal_path, method: :patch, data: { turbo_confirm: "本当に退会しますか？。この操作は取り消せません。" }, class: "btn btn-primary w-100" %>
+                    <%= button_to "退会する", withdrawal_path, method: :patch, data: {turbo_confirm: "本当に退会しますか？。この操作は取り消せません。" }, class: "btn btn-primary w-100" %>
                 </div>
             </div>
