+++ config/locales/ja.yml
@@ -3,11 +3,18 @@ ja:
     models:
       sauna: "サウナ施設"
       tweet: "投稿"
+      user: "ユーザー"
     attributes:
       tweet:
         count: "セット数"
       sauna:
         name: "名前"
+      user:
+        email: "メールアドレス"
+        name: "名前"
+        password: "パスワード"
+        password_confirmation: "確認用パスワード"
+        remember_me: "次回から自動的にログイン"
   time:
     formats:
       default: "%Y年%m月%d日 %H:%M:%S"
