+++ config/routes.rb
@@ -10,7 +10,7 @@ Rails.application.routes.draw do
   # 管理者用
   # URL /admin/sign_in ...
   devise_for :managers, controllers: {
-    registrations: "admin/registrations",
+    #registrations: "admin/registrations",
     sessions: 'admin/sessions'
   }
   
@@ -20,7 +20,7 @@ Rails.application.routes.draw do
     resources :users, only: [:show, :edit, :update]
     get "/user/unsubscribe" => "users#unsubscribe", as: 'unsubscribe'
     patch "/user/withdraw" => "users#withdrawal", as: 'withdrawal'
-    resources :saunas, only: [:show, :index, :create, :update] do
+    resources :saunas, only: [:show, :index] do
       resources :sauna_comments, only: [:create, :destroy]
       resource :sauna_favorite, only: [:create, :destroy]
       collection do
@@ -41,8 +41,8 @@ Rails.application.routes.draw do
 
   namespace :admin do
     root to: "homes#top"
-    resources :tweet, only: [:index, :show]
-    resources :saunas, except: [:update, :create]
+    resources :tweets, only: [:index, :show, :destroy]
+    resources :saunas
   end
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
